backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aiosqlite
alembic
pydantic-settings
python-jose[cryptography]
httpx
redis
celery
python-multipart
# Security
passlib[bcrypt]
# AI Pipeline
langgraph
langchain-openai
langchain-google-genai
chromadb
pyyaml
# Calendar Sync
google-auth
google-auth-oauthlib
google-api-python-client
msal
